Use conda-forge's X.org X11 headers on macOS
As per a new comment in build.sh: on macOS, we use the Aqua backend, which
causes Tk to install its own dummy-ish X11 headers. (Because Tk needs X11
declarations regardless of which backend it uses.) However, on conda-forge, we
have the X.org X11 stack, and Tk's headers are not compatible with the
rest of the stack. To prevent the packages from stomping on each other, we
explicitly remove Tk's dummy headers, and instead declare a runtime dependency
on xorg-libx11.

Note that this dependency should really be a *devel*-only dependency -- Tk
libraries are not linking with X11 libraries, so this only comes into play if
you're trying to compile something against Tk. But we don't currently have the
main/devel distinction in the packaging infrastructure.